producer/requirements.txt
atproto==0.0.55
confluent-kafka==2.5.0
python-dotenv==1.0.1
